Move number-to-string functions to libs
