+++ Marlin/src/core/utility.cpp
@@ -48,280 +48,13 @@ void safe_delay(millis_t ms) {
 
 #endif // EEPROM_SETTINGS || SD_FIRMWARE_UPDATE
 
-#if ANY(ULTRA_LCD, DEBUG_LEVELING_FEATURE, EXTENSIBLE_UI)
-
-  char conv[8] = { 0 };
-
-  #define DIGIT(n) ('0' + (n))
-  #define DIGIMOD(n, f) DIGIT((n)/(f) % 10)
-  #define RJDIGIT(n, f) ((n) >= (f) ? DIGIMOD(n, f) : ' ')
-  #define MINUSOR(n, alt) (n >= 0 ? (alt) : (n = -n, '-'))
-
-  // Convert a full-range unsigned 8bit int to a percentage
-  char* ui8tostr4pct(const uint8_t i) {
-    const uint8_t n = ui8_to_percent(i);
-    conv[3] = RJDIGIT(n, 100);
-    conv[4] = RJDIGIT(n, 10);
-    conv[5] = DIGIMOD(n, 1);
-    conv[6] = '%';
-    return &conv[3];
-  }
-
-  // Convert unsigned 8bit int to string 123 format
-  char* ui8tostr3(const uint8_t i) {
-    conv[4] = RJDIGIT(i, 100);
-    conv[5] = RJDIGIT(i, 10);
-    conv[6] = DIGIMOD(i, 1);
-    return &conv[4];
-  }
-
-  // Convert signed 8bit int to rj string with 123 or -12 format
-  char* i8tostr3(const int8_t x) {
-    int xx = x;
-    conv[4] = MINUSOR(xx, RJDIGIT(xx, 100));
-    conv[5] = RJDIGIT(xx, 10);
-    conv[6] = DIGIMOD(xx, 1);
-    return &conv[4];
-  }
-
-  // Convert unsigned 16bit int to string 123 format
-  char* ui16tostr3(const uint16_t xx) {
-    conv[4] = RJDIGIT(xx, 100);
-    conv[5] = RJDIGIT(xx, 10);
-    conv[6] = DIGIMOD(xx, 1);
-    return &conv[4];
-  }
-
-  // Convert unsigned 16bit int to string 1234 format
-  char* ui16tostr4(const uint16_t xx) {
-    conv[3] = RJDIGIT(xx, 1000);
-    conv[4] = RJDIGIT(xx, 100);
-    conv[5] = RJDIGIT(xx, 10);
-    conv[6] = DIGIMOD(xx, 1);
-    return &conv[3];
-  }
-
-  // Convert signed 16bit int to rj string with 123 or -12 format
-  char* i16tostr3(const int16_t x) {
-    int xx = x;
-    conv[4] = MINUSOR(xx, RJDIGIT(xx, 100));
-    conv[5] = RJDIGIT(xx, 10);
-    conv[6] = DIGIMOD(xx, 1);
-    return &conv[4];
-  }
-
-  // Convert unsigned 16bit int to lj string with 123 format
-  char* i16tostr3left(const int16_t i) {
-    char *str = &conv[6];
-    *str = DIGIMOD(i, 1);
-    if (i >= 10) {
-      *(--str) = DIGIMOD(i, 10);
-      if (i >= 100)
-        *(--str) = DIGIMOD(i, 100);
-    }
-    return str;
-  }
-
-  // Convert signed 16bit int to rj string with 1234, _123, -123, _-12, or __-1 format
-  char* i16tostr4sign(const int16_t i) {
-    const bool neg = i < 0;
-    const int ii = neg ? -i : i;
-    if (i >= 1000) {
-      conv[3] = DIGIMOD(ii, 1000);
-      conv[4] = DIGIMOD(ii, 100);
-      conv[5] = DIGIMOD(ii, 10);
-    }
-    else if (ii >= 100) {
-      conv[3] = neg ? '-' : ' ';
-      conv[4] = DIGIMOD(ii, 100);
-      conv[5] = DIGIMOD(ii, 10);
-    }
-    else {
-      conv[3] = ' ';
-      conv[4] = ' ';
-      if (ii >= 10) {
-        conv[4] = neg ? '-' : ' ';
-        conv[5] = DIGIMOD(ii, 10);
-      }
-      else {
-        conv[5] = neg ? '-' : ' ';
-      }
-    }
-    conv[6] = DIGIMOD(ii, 1);
-    return &conv[3];
-  }
-
-  // Convert unsigned float to string with 1.23 format
-  char* ftostr12ns(const float &f) {
-    const long i = ((f < 0 ? -f : f) * 1000 + 5) / 10;
-    conv[3] = DIGIMOD(i, 100);
-    conv[4] = '.';
-    conv[5] = DIGIMOD(i, 10);
-    conv[6] = DIGIMOD(i, 1);
-    return &conv[3];
-  }
-
-  // Convert signed float to fixed-length string with 12.34 / -2.34 format or 123.45 / -23.45 format
-  char* ftostr42_52(const float &f) {
-    if (f <= -10 || f >= 100) return ftostr52(f); // need more digits
-    long i = (f * 1000 + (f < 0 ? -5: 5)) / 10;
-    conv[2] = (f >= 0 && f < 10) ? ' ' : MINUSOR(i, DIGIMOD(i, 1000));
-    conv[3] = DIGIMOD(i, 100);
-    conv[4] = '.';
-    conv[5] = DIGIMOD(i, 10);
-    conv[6] = DIGIMOD(i, 1);
-    return &conv[2];
-  }
-
-  // Convert signed float to fixed-length string with 023.45 / -23.45 format
-  char* ftostr52(const float &f) {
-    long i = (f * 1000 + (f < 0 ? -5: 5)) / 10;
-    conv[1] = MINUSOR(i, DIGIMOD(i, 10000));
-    conv[2] = DIGIMOD(i, 1000);
-    conv[3] = DIGIMOD(i, 100);
-    conv[4] = '.';
-    conv[5] = DIGIMOD(i, 10);
-    conv[6] = DIGIMOD(i, 1);
-    return &conv[1];
-  }
-
-  #if ENABLED(LCD_DECIMAL_SMALL_XY)
-
-    // Convert float to rj string with 1234, _123, -123, _-12, 12.3, _1.2, or -1.2 format
-    char* ftostr4sign(const float &f) {
-      const int i = (f * 100 + (f < 0 ? -5: 5)) / 10;
-      if (!WITHIN(i, -99, 999)) return i16tostr4sign((int)f);
-      const bool neg = i < 0;
-      const int ii = neg ? -i : i;
-      conv[3] = neg ? '-' : (ii >= 100 ? DIGIMOD(ii, 100) : ' ');
-      conv[4] = DIGIMOD(ii, 10);
-      conv[5] = '.';
-      conv[6] = DIGIMOD(ii, 1);
-      return &conv[3];
-    }
-
-  #endif // LCD_DECIMAL_SMALL_XY
-
-  // Convert float to fixed-length string with +123.4 / -123.4 format
-  char* ftostr41sign(const float &f) {
-    int i = (f * 100 + (f < 0 ? -5: 5)) / 10;
-    conv[1] = MINUSOR(i, '+');
-    conv[2] = DIGIMOD(i, 1000);
-    conv[3] = DIGIMOD(i, 100);
-    conv[4] = DIGIMOD(i, 10);
-    conv[5] = '.';
-    conv[6] = DIGIMOD(i, 1);
-    return &conv[1];
-  }
-
-  // Convert signed float to string (6 digit) with -1.234 / _0.000 / +1.234 format
-  char* ftostr43sign(const float &f, char plus/*=' '*/) {
-    long i = (f * 10000 + (f < 0 ? -5: 5)) / 10;
-    conv[1] = i ? MINUSOR(i, plus) : ' ';
-    conv[2] = DIGIMOD(i, 1000);
-    conv[3] = '.';
-    conv[4] = DIGIMOD(i, 100);
-    conv[5] = DIGIMOD(i, 10);
-    conv[6] = DIGIMOD(i, 1);
-    return &conv[1];
-  }
-
-  // Convert signed float to string (5 digit) with -1.2345 / _0.0000 / +1.2345 format
-  char* ftostr54sign(const float &f, char plus/*=' '*/) {
-    long i = (f * 100000 + (f < 0 ? -5: 5)) / 10;
-    conv[0] = i ? MINUSOR(i, plus) : ' ';
-    conv[1] = DIGIMOD(i, 10000);
-    conv[2] = '.';
-    conv[3] = DIGIMOD(i, 1000);
-    conv[4] = DIGIMOD(i, 100);
-    conv[5] = DIGIMOD(i, 10);
-    conv[6] = DIGIMOD(i, 1);
-    return &conv[0];
-  }
-
-  // Convert unsigned float to rj string with 12345 format
-  char* ftostr5rj(const float &f) {
-    const long i = ((f < 0 ? -f : f) * 10 + 5) / 10;
-    conv[2] = RJDIGIT(i, 10000);
-    conv[3] = RJDIGIT(i, 1000);
-    conv[4] = RJDIGIT(i, 100);
-    conv[5] = RJDIGIT(i, 10);
-    conv[6] = DIGIMOD(i, 1);
-    return &conv[2];
-  }
-
-  // Convert signed float to string with +1234.5 format
-  char* ftostr51sign(const float &f) {
-    long i = (f * 100 + (f < 0 ? -5: 5)) / 10;
-    conv[0] = MINUSOR(i, '+');
-    conv[1] = DIGIMOD(i, 10000);
-    conv[2] = DIGIMOD(i, 1000);
-    conv[3] = DIGIMOD(i, 100);
-    conv[4] = DIGIMOD(i, 10);
-    conv[5] = '.';
-    conv[6] = DIGIMOD(i, 1);
-    return conv;
-  }
-
-  // Convert signed float to string with +123.45 format
-  char* ftostr52sign(const float &f) {
-    long i = (f * 1000 + (f < 0 ? -5: 5)) / 10;
-    conv[0] = MINUSOR(i, '+');
-    conv[1] = DIGIMOD(i, 10000);
-    conv[2] = DIGIMOD(i, 1000);
-    conv[3] = DIGIMOD(i, 100);
-    conv[4] = '.';
-    conv[5] = DIGIMOD(i, 10);
-    conv[6] = DIGIMOD(i, 1);
-    return conv;
-  }
-
-  // Convert unsigned float to string with 1234.5 format omitting trailing zeros
-  char* ftostr51rj(const float &f) {
-    const long i = ((f < 0 ? -f : f) * 100 + 5) / 10;
-    conv[0] = ' ';
-    conv[1] = RJDIGIT(i, 10000);
-    conv[2] = RJDIGIT(i, 1000);
-    conv[3] = RJDIGIT(i, 100);
-    conv[4] = DIGIMOD(i, 10);
-    conv[5] = '.';
-    conv[6] = DIGIMOD(i, 1);
-    return conv;
-  }
-
-  // Convert signed float to space-padded string with -_23.4_ format
-  char* ftostr52sp(const float &f) {
-    long i = (f * 1000 + (f < 0 ? -5: 5)) / 10;
-    uint8_t dig;
-    conv[0] = MINUSOR(i, ' ');
-    conv[1] = RJDIGIT(i, 10000);
-    conv[2] = RJDIGIT(i, 1000);
-    conv[3] = DIGIMOD(i, 100);
-
-    if ((dig = i % 10)) {          // second digit after decimal point?
-      conv[4] = '.';
-      conv[5] = DIGIMOD(i, 10);
-      conv[6] = DIGIT(dig);
-    }
-    else {
-      if ((dig = (i / 10) % 10)) { // first digit after decimal point?
-        conv[4] = '.';
-        conv[5] = DIGIT(dig);
-      }
-      else                          // nothing after decimal point
-        conv[4] = conv[5] = ' ';
-      conv[6] = ' ';
-    }
-    return conv;
-  }
-
-#endif // ULTRA_LCD
-
 #if ENABLED(DEBUG_LEVELING_FEATURE)
 
   #include "../module/probe.h"
   #include "../module/motion.h"
   #include "../module/stepper.h"
+  #include "../module/stepper.h"
+  #include "../libs/numtostr.h"
   #include "../feature/bedlevel/bedlevel.h"
 
   void log_machine_info() {
