+++ Marlin/src/lcd/dogm/status_screen_DOGM.cpp
@@ -33,6 +33,7 @@
 #include "ultralcd_DOGM.h"
 #include "../ultralcd.h"
 #include "../lcdprint.h"
+#include "../../libs/numtostr.h"
 
 #include "../../module/motion.h"
 #include "../../module/temperature.h"
